Move HTTPConnectionProtocol to http.client.
This keeps it next to HTTPConnection where it is maintainable.
