+++ stdlib/3/urllib/request.pyi
@@ -2,9 +2,9 @@
 
 from typing import (
     Any, Callable, ClassVar, Dict, List, IO, Mapping, Optional, Sequence, Tuple,
-    TypeVar, Union, overload, NoReturn, Protocol,
+    TypeVar, Union, overload, NoReturn,
 )
-from http.client import HTTPResponse, HTTPMessage
+from http.client import HTTPResponse, HTTPMessage, HTTPConnectionProtocol
 from http.cookiejar import CookieJar
 from email.message import Message
 from urllib.response import addinfourl
@@ -157,22 +157,10 @@ class ProxyDigestAuthHandler(BaseHandler, AbstractDigestAuthHandler):
     def http_error_407(self, req: Request, fp: IO[str], code: int, msg: int,
                        hdrs: Mapping[str, str]) -> Optional[_UrlopenRet]: ...
 
-# TODO: Could this just be turned into something importing from http.client?
-class _HTTPConnectionProtocol(Protocol):  # http.client.HTTPConnection
-    if sys.version_info >= (3, 7):
-        def __call__(self, host: str, port: Optional[int] = ...,
-                     timeout: int = ...,
-                     source_address: Optional[Tuple[str, int]] = ...,
-                     blocksize: int = ...): ...
-    else:
-        def __call__(self, host: str, port: Optional[int] = ...,
-                     timeout: int = ...,
-                     source_address: Optional[Tuple[str, int]] = ...): ...
-
 class HTTPHandler(BaseHandler):
     def http_open(self, req: Request) -> HTTPResponse: ...
     def do_open(self,  # undocumented
-                http_class: _HTTPConnectionProtocol,
+                http_class: HTTPConnectionProtocol,
                 req: Request) -> HTTPResponse: ...
 
 class HTTPSHandler(BaseHandler):
